+++ MuxExoPlayer/src/main/java/com/mux/stats/sdk/muxstats/AdsImaSDKListener.java
@@ -19,6 +19,7 @@ public class AdsImaSDKListener implements AdErrorEvent.AdErrorListener, AdEvent.
 
   private final MuxBaseExoPlayer exoPlayerListener;
   private boolean sendPlayOnStarted = false;
+  private boolean missingAdBreakStartEvent = false;
 
   public AdsImaSDKListener(MuxBaseExoPlayer listener) {
     exoPlayerListener = listener;
@@ -61,10 +62,17 @@ public class AdsImaSDKListener implements AdErrorEvent.AdErrorListener, AdEvent.
               exoPlayerListener.getState() == MuxBaseExoPlayer.PlayerState.PLAYING) {
             exoPlayerListener.pause();
           }
+          sendPlayOnStarted = false;
+          exoPlayerListener.setState(MuxBaseExoPlayer.PlayerState.PLAYING_ADS);
+          if (!exoPlayerListener.player.get().getPlayWhenReady() &&
+              (exoPlayerListener.player.get().getCurrentPosition() == 0)) {
+            // This is preroll ads when play when ready is set to false, we need to ignore these events
+            missingAdBreakStartEvent = true;
+            break;
+          }
           exoPlayerListener.setState(MuxBaseExoPlayer.PlayerState.PLAYING_ADS);
           dispatchAdPlaybackEvent(new AdBreakStartEvent(null), ad);
           dispatchAdPlaybackEvent(new AdPlayEvent(null), ad);
-          sendPlayOnStarted = false;
           break;
         case STARTED:
           // On the first STARTED, do not send AdPlay, as it was handled in
@@ -99,11 +107,24 @@ public class AdsImaSDKListener implements AdErrorEvent.AdErrorListener, AdEvent.
           }
           break;
         case PAUSED:
+          if (!exoPlayerListener.player.get().getPlayWhenReady() &&
+              (exoPlayerListener.player.get().getCurrentPosition() == 0)) {
+            // This is preroll ads when play when ready is set to false, we need to ignore these events
+            break;
+          }
           dispatchAdPlaybackEvent(new AdPauseEvent(null), ad);
           break;
         case RESUMED:
-          dispatchAdPlaybackEvent(new AdPlayEvent(null), ad);
-          dispatchAdPlaybackEvent(new AdPlayingEvent(null), ad);
+          if (missingAdBreakStartEvent) {
+            // This is special case when we have ad preroll and play when ready is set to false
+            // in that case we need to dispatch AdBreakStartEvent first and resume the playback.
+            dispatchAdPlaybackEvent(new AdBreakStartEvent(null), ad);
+            dispatchAdPlaybackEvent(new AdPlayEvent(null), ad);
+            missingAdBreakStartEvent = false;
+          } else {
+            dispatchAdPlaybackEvent(new AdPlayEvent(null), ad);
+            dispatchAdPlaybackEvent(new AdPlayingEvent(null), ad);
+          }
           break;
         case ALL_ADS_COMPLETED:
           // Nothing to do here, as this depends on VAST vs VMAP and is not