+++ MuxExoPlayer/src/main/java/com/mux/stats/sdk/muxstats/MuxBaseExoPlayer.java
@@ -87,6 +87,10 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
   protected WeakReference<View> playerView;
   protected WeakReference<Context> contextRef;
   protected AdsImaSDKListener adsImaSdkListener;
+
+  protected int numberOfEventsSent = 0;
+  protected int numberOfPlayEventsSent = 0;
+  protected int numberOfPauseEventsSent = 0;
   protected int streamType = -1;
 
   public enum PlayerState {
@@ -244,11 +248,15 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
 
   @SuppressWarnings("unused")
   public void videoChange(CustomerVideoData customerVideoData) {
+    // Reset the state to avoid unwanted rebuffering events
+    state = PlayerState.INIT;
+    resetInternalStats();
     muxStats.videoChange(customerVideoData);
   }
 
   @SuppressWarnings("unused")
   public void programChange(CustomerVideoData customerVideoData) {
+    resetInternalStats();
     muxStats.programChange(customerVideoData);
   }
 
@@ -292,6 +300,13 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
   @Override
   public void dispatch(IEvent event) {
     if (player != null && player.get() != null && muxStats != null) {
+      numberOfEventsSent++;
+      if (event.getType().equalsIgnoreCase(PlayEvent.TYPE)) {
+        numberOfPlayEventsSent++;
+      }
+      if (event.getType().equalsIgnoreCase(PauseEvent.TYPE)) {
+        numberOfPauseEventsSent++;
+      }
       super.dispatch(event);
     }
   }
@@ -439,7 +454,7 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
   }
 
   protected void pause() {
-    if (state == PlayerState.SEEKED) {
+    if (state == PlayerState.SEEKED && numberOfPauseEventsSent > 0) {
       // No pause event after seeked
       return;
     }
@@ -455,9 +470,14 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
   }
 
   protected void play() {
-    if (state == PlayerState.REBUFFERING
-        || seekingInProgress
-        || state == PlayerState.SEEKED) {
+    // If this is the first play event it may be very important not to be skipped
+    // In all other cases skip this play event
+    if (
+        (state == PlayerState.REBUFFERING
+            || seekingInProgress
+            || state == PlayerState.SEEKED) &&
+            (numberOfPlayEventsSent > 0)
+    ) {
       // Ignore play event after rebuffering and Seeking
       return;
     }
@@ -553,6 +573,12 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
       RenditionChangeEvent event = new RenditionChangeEvent(null);
       dispatch(event);
     }
+  }
+
+  private void resetInternalStats() {
+    numberOfPauseEventsSent = 0;
+    numberOfPlayEventsSent = 0;
+    numberOfEventsSent = 0;
   }
 
   static class FrameRenderedListener implements VideoFrameMetadataListener {